[ASN-144] fix: float로 통일

+++ apps/core-api/src/main/java/com/ott/core/modules/recommendation/service/RecommendationService.java
@@ -60,7 +60,7 @@ public class RecommendationService {
     // =========================================================================
     public List<VideoFeedResponseDto> getPersonalizedFeed(Long userId, VideoType videoType, int page, int size) {
 
-        List<Double> userVector = userVectorService.getUserVector(userId);
+        List<Float> userVector = userVectorService.getUserVector(userId);
 
         NativeQuery searchQuery;
         if (userVector == null || userVector.isEmpty()) {
